+++ CHANGES.md
@@ -2,6 +2,7 @@
 # Unreleased
 
 * use `GDAL_MEM_ENABLE_OPEN=TRUE` when opening a numpy array with rasterio
+* add `interpolate=True/False` to `.point()` methods to allow interpolation of surrounding pixels
 
 # 7.5.0 (2025-02-26)